requirements.txt
Flask==2.0.1
psycopg2-binary==2.9.1
gunicorn==20.1.0
werkzeug==2.0.1
fastapi>=0.68.1
uvicorn>=0.15.0
httpx>=0.21.0
requests>=2.26.0
annotated-types>=0.4.0
pydantic-core>=2.14.5
typing-extensions>=4.6.1